Split logic into hooks.
